requirements.txt
flask
gunicorn
requests
pydub
boto3==1.38.35
openai>=1.14.2
psutil
gevent==25.5.1
apscheduler
google-cloud-vision
notion-client>=2.2
concurrent-log-handler>=0.9
SQLAlchemy>=1.4,<2.0
psycopg2-binary>=2.9,<2.10
pytest
python-dotenv